+++ apollo.config.js
@@ -6,7 +6,6 @@ module.exports = {
     excludes: ["**/*.test.ts", "**/__tests__/*"]
   },
   engine: {
-    frontend: "https://engine-staging.apollographql.com",
     endpoint: "https://engine-staging-graphql.apollographql.com/api/graphql"
   }
 };